change folder structure to reduce the size of gym_manip
